backend/requirements.txt
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Flask==3.1.2
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-vision==3.11.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
memory-profiler==0.61.0
numpy==1.25.2
packaging==25.0
Pillow==9.5.0
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pytesseract==0.3.10
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
zipp==3.23.0
